Update createPptx.R
if 2 comparisons have overlapping characters in their names, for example "MAB_over_MAC" and "MAB_over_MA", this was causing issues with adding pcas to the summary pptx because the grepl was picking up both the comparisons when looking for "MAB_over_MA". Fixed this by using comparison labels to add the pcas to the summary pptx. closes #44