Update createPptx.R
During summary pptx generation, pipeline was initially not taking into account multiple methods or sheets in the input data and hence failing to generate the summary pptx. issue resolved by making that piece of code dynamic to take into account the number of methods or sheets in the input data.